+++ test/calculate_test.go
@@ -4,7 +4,7 @@ import "testing"
 import "../cmd/calculate"
 
 func TestCalculateDivision(t *testing.T) {
-	expected := float64(7.5)
+	expected := float32(7.5)
 	actual, err := calculate.Calculate(15, 2, "DIVIDE")
 	if actual != expected {
 		t.Error("Test failed")
@@ -15,7 +15,7 @@ func TestCalculateDivision(t *testing.T) {
 }
 
 func TestCalculateMultiply(t *testing.T) {
-	expected := float64(30)
+	expected := float32(30)
 	actual, err := calculate.Calculate(15, 2, "MULTIPLY")
 	if actual != expected {
 		t.Error("Test failed")
@@ -26,7 +26,7 @@ func TestCalculateMultiply(t *testing.T) {
 }
 
 func TestCalculateAdd(t *testing.T) {
-	expected := float64(3)
+	expected := float32(3)
 	actual, err := calculate.Calculate(1, 2, "ADD")
 	if actual != expected {
 		t.Error("Test failed")
@@ -37,7 +37,7 @@ func TestCalculateAdd(t *testing.T) {
 }
 
 func TestCalculateSubtract(t *testing.T) {
-	expected := float64(-1)
+	expected := float32(-1)
 	actual, err := calculate.Calculate(1, 2, "SUBTRACT")
 	if actual != expected {
 		t.Error("Test failed")
